+++ components/EN/Footer.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+import { AiOutlineGithub, AiOutlineLinkedin } from "react-icons/ai";
+
+export default function Footer() {
+  return (
+    <footer className="mx-auto max-w-3xl px-4 sm:px-6 md:max-w-5xl ">
+      <hr className="w-full h-0.5 mx-auto mt-8 bg-neutral-200 border-0" />
+      <div className="mx-auto  p-4 flex flex-col text-center text-neutral-900 md:flex-row md:justify-between">
+        <div className="flex flex-row items-center justify-center space-x-1 text-neutral-500 dark:text-neutral-100">
+          © 2023 Gabriel Henrique
+          <a className="hover:underline" />
+        </div>
+        <div className="flex flex-row items-center justify-center space-x-2 mb-1">
+          <a
+            href="https://github.com/Gabrielhsb"
+            rel="noreferrer"
+            target="_blank"
+          >
+            <AiOutlineGithub
+              className="hover:-translate-y-1 transition-transform cursor-pointer text-neutral-500 dark:text-neutral-100"
+              size={30}
+            />
+          </a>
+          <a
+            href="https://www.linkedin.com/in/gabriel-bernardo-75a436b7/"
+            rel="noreferrer"
+            target="_blank"
+          >
+            <AiOutlineLinkedin
+              className="hover:-translate-y-1 transition-transform cursor-pointer text-neutral-500 dark:text-neutral-100"
+              size={30}
+            />
+          </a>
+        </div>
+      </div>
+    </footer>
+  );
+}
